Create README.md
Hey, I'm Roetzee. I wanted to share my Rust cheat that I worked on a while back so you can take something from it or even use it as a base for something new. This is an internal cheat. Unfortunately, I deleted my Loader for it a long time ago, so I won't be able to share it. The cheat was made for version 2511. It features automatic offset updates and almost perfectly tuned math. Enjoy.

And goodbye, Expensive Rust DLC.

If you have any questions, feel free to reach out:
Discord: Roetzee
Telegram: @Roetzee